e2e tests
- competition tests
- handling of eventual consistency race condition
