+++ apps/hms-api/test/app.e2e-spec.ts
@@ -1,3 +1,4 @@
+// eslint-disable-next-line max-classes-per-file
 import { EventStoreDBClient } from '@eventstore/db-client';
 import { ApolloDriver, ApolloDriverConfig } from '@nestjs/apollo';
 import { ExecutionContext, INestApplication, Injectable } from '@nestjs/common';
@@ -5,7 +6,7 @@ import { ConfigModule } from '@nestjs/config';
 import { CommandBus, CqrsModule, QueryBus } from '@nestjs/cqrs';
 import { GraphQLModule } from '@nestjs/graphql';
 import { Test, TestingModule } from '@nestjs/testing';
-import supertest, { Request } from 'supertest';
+import { Request } from 'supertest';
 import { ApplicationModule } from '../src/modules/application/application.module';
 import { ClassificationModule } from '../src/modules/classification/classification.module';
 import { CommentModule } from '../src/modules/comment/comment.module';
@@ -26,6 +27,36 @@ import { SkillModule } from '../src/modules/skill/skill.module';
 import { SyncUserCommand } from '../src/modules/user/commands/sync-user/sync-user.command';
 import { UserModule } from '../src/modules/user/user.module';
 import { applySeeds, seedsExist } from '../src/seeds';
+import { competitionTests } from './competition-tests';
+import { seedTests } from './seed-tests';
+
+// seeding takes a while
+jest.setTimeout(90000);
+
+export const testContext = {
+  app: null,
+};
+
+export async function pollUntilTrue(
+  testFunction: () => Promise<boolean>,
+  timeout = 5000,
+  interval = 100,
+): Promise<void> {
+  // Since SQRS provides eventual consistency, making mutations does not immidietely propagate results to the read db
+  const startTime = Date.now();
+
+  while (Date.now() - startTime < timeout) {
+    // eslint-disable-next-line no-await-in-loop
+    const result = await testFunction();
+    if (result) {
+      return;
+    }
+    // eslint-disable-next-line no-await-in-loop, no-promise-executor-return
+    await new Promise((res) => setTimeout(res, interval));
+  }
+
+  throw new Error('Polling timed out');
+}
 
 @Injectable()
 export class MockAuthGuard {
@@ -37,9 +68,6 @@ export class MockAuthGuard {
     return true;
   }
 }
-
-// seeding takes a while
-jest.setTimeout(90000);
 
 describe('E2E tests', () => {
   // let resolver: CompetitionResolver;
@@ -90,6 +118,7 @@ describe('E2E tests', () => {
     // resolver = module.get<CompetitionResolver>(CompetitionResolver);
     app = module.createNestApplication<INestApplication>();
 
+    // attach a fake user to every request
     app.use((req, res, next) => {
       req.user = {
         id: '4e0b74a8-1b63-47fa-a082-684ab7301ea9',
@@ -103,6 +132,7 @@ describe('E2E tests', () => {
 
     await app.init();
 
+    // apply seeds if they haven't been applied before
     if (!(await seedsExist(queryBus))) {
       await applySeeds(commandBus, eventStore);
     }
@@ -130,588 +160,11 @@ describe('E2E tests', () => {
     // // Print the schema
     // const printedSchema = printSchema(schema);
     // console.log(printedSchema);
+    testContext.app = app;
   });
 
-  describe('Seed checks', () => {
-    it(`should have system admin user and it is the only user`, async () => {
-      const query = `
-      query {
-        users {
-          id
-          deltek_id
-          name
-          email
-          roles
-          created_at
-          updated_at
-        }
-      }
-    `;
-
-      return supertest(app.getHttpServer())
-        .post('/graphql')
-        .send({ query })
-        .expect(200)
-        .then((response) => {
-          // Check that users are defined and an array
-          expect(response.body.data.users).toBeDefined();
-          expect(Array.isArray(response.body.data.users)).toBe(true);
-
-          // Check there is only one user
-          expect(response.body.data.users.length).toEqual(1);
-
-          // Get the first user from the array
-          const [user] = response.body.data.users;
-
-          // Check the details of that user
-          expect(user.name).toEqual('System User');
-          expect(user.roles).toEqual(expect.arrayContaining(['admin']));
-        });
-    });
-
-    it(`should have correct number of grid seeds`, async () => {
-      const query = `
-      query {
-        grids {
-          id
-        }
-      }
-    `;
-
-      return supertest(app.getHttpServer())
-        .post('/graphql')
-        .send({ query })
-        .expect(200)
-        .then((response) => {
-          // Check that users are defined and an array
-          expect(response.body.data.grids).toBeDefined();
-          expect(Array.isArray(response.body.data.grids)).toBe(true);
-
-          // Check there is only one user
-          expect(response.body.data.grids.length).toEqual(6);
-        });
-    });
-
-    it(`should have correct number of occupation group seeds`, async () => {
-      const query = `
-      query {
-        occupationGroups {
-          id
-        }
-      }
-    `;
-
-      return supertest(app.getHttpServer())
-        .post('/graphql')
-        .send({ query })
-        .expect(200)
-        .then((response) => {
-          // Check that users are defined and an array
-          expect(response.body.data.occupationGroups).toBeDefined();
-          expect(Array.isArray(response.body.data.occupationGroups)).toBe(true);
-
-          // Check there is only one user
-          expect(response.body.data.occupationGroups.length).toEqual(3);
-        });
-    });
-
-    it(`should have correct number of classifications seeds`, async () => {
-      const query = `
-      query {
-        classifications {
-          id
-        }
-      }
-    `;
-
-      return supertest(app.getHttpServer())
-        .post('/graphql')
-        .send({ query })
-        .expect(200)
-        .then((response) => {
-          // Check that users are defined and an array
-          expect(response.body.data.classifications).toBeDefined();
-          expect(Array.isArray(response.body.data.classifications)).toBe(true);
-
-          // Check there is only one user
-          expect(response.body.data.classifications.length).toEqual(8);
-        });
-    });
-
-    it(`should have correct number of jobDescriptions seeds`, async () => {
-      const query = `
-      query {
-        jobDescriptions {
-          id
-        }
-      }
-    `;
-
-      return supertest(app.getHttpServer())
-        .post('/graphql')
-        .send({ query })
-        .expect(200)
-        .then((response) => {
-          // Check that users are defined and an array
-          expect(response.body.data.jobDescriptions).toBeDefined();
-          expect(Array.isArray(response.body.data.jobDescriptions)).toBe(true);
-
-          // Check there is only one user
-          expect(response.body.data.jobDescriptions.length).toEqual(16);
-        });
-    });
-
-    it(`should have correct number of ministries seeds`, async () => {
-      const query = `
-      query {
-        ministries {
-          id
-        }
-      }
-    `;
-
-      return supertest(app.getHttpServer())
-        .post('/graphql')
-        .send({ query })
-        .expect(200)
-        .then((response) => {
-          // Check that users are defined and an array
-          expect(response.body.data.ministries).toBeDefined();
-          expect(Array.isArray(response.body.data.ministries)).toBe(true);
-
-          // Check there is only one user
-          expect(response.body.data.ministries.length).toEqual(47);
-        });
-    });
-
-    it(`should have correct number of skills seeds`, async () => {
-      const query = `
-      query {
-        skills {
-          id
-        }
-      }
-    `;
-
-      return supertest(app.getHttpServer())
-        .post('/graphql')
-        .send({ query })
-        .expect(200)
-        .then((response) => {
-          // Check that users are defined and an array
-          expect(response.body.data.skills).toBeDefined();
-          expect(Array.isArray(response.body.data.skills)).toBe(true);
-
-          // Check there is only one user
-          expect(response.body.data.skills.length).toEqual(444);
-        });
-    });
-
-    it(`should have correct number of locations seeds`, async () => {
-      const query = `
-      query {
-        locations {
-          id
-        }
-      }
-    `;
-
-      return supertest(app.getHttpServer())
-        .post('/graphql')
-        .send({ query })
-        .expect(200)
-        .then((response) => {
-          // Check that users are defined and an array
-          expect(response.body.data.locations).toBeDefined();
-          expect(Array.isArray(response.body.data.locations)).toBe(true);
-
-          // Check there is only one user
-          expect(response.body.data.locations.length).toEqual(168);
-        });
-    });
-  });
-
-  describe('Competition tests', () => {
-    it(`should create competition`, async () => {
-      const postData = {
-        query: `
-      mutation CreateCompetition($data: CreateCompetitionInput!) {
-        createCompetition(data: $data)
-      }`,
-
-        variables: {
-          data: {
-            id: 'fb0b9943-c133-4879-8ff5-1a3c204ed669',
-            job_description_id: '09ab8e7d-9cc1-4993-8d53-14095dd00e81',
-            recruiter_id: '4e0b74a8-1b63-47fa-a082-684ab7301ea9',
-            category: 'CMH',
-            deltek_id: null,
-          },
-        },
-      };
-
-      return supertest(app.getHttpServer())
-        .post('/graphql')
-        .send(postData)
-        .expect(200)
-        .then((response) => {
-          // Check if data exists in the response body
-          expect(response.body.data).toBeDefined();
-
-          // Check if createCompetition exists in the data
-          expect(response.body.data.createCompetition).toBeDefined();
-
-          // Check if the returned id is correct
-          expect(response.body.data.createCompetition).toEqual('fb0b9943-c133-4879-8ff5-1a3c204ed669');
-        });
-    });
-
-    it(`should fail creating competition with existing id`, async () => {
-      const postData = {
-        query: `
-      mutation CreateCompetition($data: CreateCompetitionInput!) {
-        createCompetition(data: $data)
-      }`,
-
-        variables: {
-          data: {
-            id: 'fb0b9943-c133-4879-8ff5-1a3c204ed669',
-            job_description_id: '09ab8e7d-9cc1-4993-8d53-14095dd00e81',
-            recruiter_id: '4e0b74a8-1b63-47fa-a082-684ab7301ea9',
-            category: 'CMH',
-            deltek_id: null,
-          },
-        },
-      };
-
-      return supertest(app.getHttpServer())
-        .post('/graphql')
-        .send(postData)
-        .expect(200)
-        .then((response) => {
-          // console.log('create response: ', response.body);
-
-          // Check if errors exist in the response body
-          expect(response.body.errors).toBeDefined();
-
-          // Check if errors array has at least one item
-          expect(response.body.errors.length).toBeGreaterThan(0);
-
-          // Check if the error message is what you expect
-          expect(response.body.errors[0].message).toEqual('Competition already exists');
-
-          // Check if data is null
-          expect(response.body.data).toBeNull();
-        });
-    });
-
-    it(`should get competition by id`, async () => {
-      // make another competition
-
-      const postData = {
-        query: `
-      mutation CreateCompetition($data: CreateCompetitionInput!) {
-        createCompetition(data: $data)
-      }`,
-
-        variables: {
-          data: {
-            id: 'fb0b9943-c133-4879-8ff5-1a3c204ed661',
-            job_description_id: '09ab8e7d-9cc1-4993-8d53-14095dd00e81',
-            recruiter_id: '4e0b74a8-1b63-47fa-a082-684ab7301ea9',
-            category: 'CMH',
-            deltek_id: null,
-          },
-        },
-      };
-
-      await supertest(app.getHttpServer())
-        .post('/graphql')
-        .send(postData)
-        .expect(200)
-        .then((response) => {
-          // // Check if data exists in the response body
-          // expect(response.body.data).toBeDefined();
-          // // Check if createCompetition exists in the data
-          // expect(response.body.data.createCompetition).toBeDefined();
-          // // Check if the returned id is correct
-          // expect(response.body.data.createCompetition).toEqual('fb0b9943-c133-4879-8ff5-1a3c204ed661');
-        });
-
-      // retreive competition
-      const query = `
-    query Competition {
-      competition(id: "fb0b9943-c133-4879-8ff5-1a3c204ed661") {
-          id
-          category
-          state
-          deltek_id
-          created_at
-          updated_at
-          deleted_at
-      }
-    } `;
-
-      await supertest(app.getHttpServer())
-        .post('/graphql')
-        .send({ query })
-        .expect(200)
-        .then((response) => {
-          // Check if 'data' and 'competition' are defined
-          expect(response.body.data).toBeDefined();
-          expect(response.body.data.competition).toBeDefined();
-
-          // Check if 'id' matches the expected value
-          expect(response.body.data.competition.id).toEqual('fb0b9943-c133-4879-8ff5-1a3c204ed661');
-
-          // Check if 'category' is "RH"
-          expect(response.body.data.competition.category).toEqual('CMH');
-
-          // Check if 'state' is "DRAFT"
-          expect(response.body.data.competition.state).toEqual('DRAFT');
-
-          // Check if 'deltek_id' is "test_deltek_id"
-          expect(response.body.data.competition.deltek_id).toBeNull();
-
-          // Check if 'created_at' and 'updated_at' are defined and are not null
-          expect(response.body.data.competition.created_at).toBeDefined();
-          expect(response.body.data.competition.updated_at).toBeNull();
-
-          // Check if 'deleted_at' is null
-          expect(response.body.data.competition.deleted_at).toBeNull();
-        });
-    });
-
-    it(`should get all competitions`, async () => {
-      // retreive competition
-      const query = `
-    query Competitions {
-      competitions {
-          id
-          category
-          state
-          deltek_id
-          created_at
-          updated_at
-          deleted_at
-      }
-    }
-    `;
-
-      await supertest(app.getHttpServer())
-        .post('/graphql')
-        .send({ query })
-        .expect(200)
-        .then((response) => {
-          // Check if 'data' and 'competition' are defined
-          expect(response.body.data).toBeDefined();
-          expect(response.body.data.competitions).toBeDefined();
-          expect(response.body.data.competitions.length).toEqual(2);
-        });
-    });
-
-    it(`should update an existing competition`, async () => {
-      const postData = {
-        query: `
-      mutation UpdateCompetition($data: UpdateCompetitionInput!) {
-        updateCompetition(data: $data)
-      }`,
-
-        variables: {
-          data: {
-            id: 'fb0b9943-c133-4879-8ff5-1a3c204ed669',
-            job_description_id: '193670cc-c74b-4fe3-9158-8175e907e29f',
-            recruiter_id: '4e0b74a8-1b63-47fa-a082-684ab7301ea0',
-            category: 'RH',
-            deltek_id: 'test_deltek_id',
-          },
-        },
-      };
-
-      await supertest(app.getHttpServer())
-        .post('/graphql')
-        .send(postData)
-        .expect(200)
-        .then((response) => {
-          // console.log('update response: ', response.body);
-
-          // Check if data exists in the response body
-          expect(response.body.data).toBeDefined();
-
-          // Check if createCompetition exists in the data
-          expect(response.body.data.updateCompetition).toBeDefined();
-
-          // Check if the returned id is correct
-          expect(response.body.data.updateCompetition).toEqual('fb0b9943-c133-4879-8ff5-1a3c204ed669');
-        });
-
-      const query = `
-    query Competition {
-      competition(id: "fb0b9943-c133-4879-8ff5-1a3c204ed669") {
-          id
-          category
-          state
-          deltek_id
-          created_at
-          updated_at
-          deleted_at
-      }
-    } `;
-
-      await supertest(app.getHttpServer())
-        .post('/graphql')
-        .send({ query })
-        .expect(200)
-        .then((response) => {
-          // Check if 'data' and 'competition' are defined
-          expect(response.body.data).toBeDefined();
-          expect(response.body.data.competition).toBeDefined();
-
-          // Check if 'id' matches the expected value
-          expect(response.body.data.competition.id).toEqual('fb0b9943-c133-4879-8ff5-1a3c204ed669');
-
-          // Check if 'category' is "RH"
-          expect(response.body.data.competition.category).toEqual('RH');
-
-          // Check if 'state' is "DRAFT"
-          expect(response.body.data.competition.state).toEqual('DRAFT');
-
-          // Check if 'deltek_id' is "test_deltek_id"
-          expect(response.body.data.competition.deltek_id).toEqual('test_deltek_id');
-
-          // Check if 'created_at' and 'updated_at' are defined and are not null
-          expect(response.body.data.competition.created_at).toBeDefined();
-          expect(response.body.data.competition.updated_at).toBeDefined();
-
-          // Check if 'deleted_at' is null
-          expect(response.body.data.competition.deleted_at).toBeNull();
-        });
-    });
-
-    it(`should fail updating competition with a non-existing id`, async () => {
-      const postData = {
-        query: `
-      mutation UpdateCompetition($data: UpdateCompetitionInput!) {
-        updateCompetition(data: $data)
-      }`,
-
-        variables: {
-          data: {
-            id: 'fb0b9943-c133-4879-8ff5-1a3c204ed660',
-            job_description_id: '193670cc-c74b-4fe3-9158-8175e907e29f',
-            recruiter_id: '4e0b74a8-1b63-47fa-a082-684ab7301ea0',
-            category: 'RH',
-            deltek_id: 'test_deltek_id',
-          },
-        },
-      };
-
-      return supertest(app.getHttpServer())
-        .post('/graphql')
-        .send(postData)
-        .expect(200)
-        .then((response) => {
-          // console.log('create response: ', response.body);
-
-          // Check if errors exist in the response body
-          expect(response.body.errors).toBeDefined();
-
-          // Check if errors array has at least one item
-          expect(response.body.errors.length).toBeGreaterThan(0);
-
-          // Check if the error message is what you expect
-          expect(response.body.errors[0].message).toEqual("Competition doesn't exist");
-
-          // Check if data is null
-          expect(response.body.data).toBeNull();
-        });
-    });
-
-    it(`should delete competition`, async () => {
-      const postData = {
-        query: `
-      mutation DeleteCompetition ($data: DeleteCompetitionInput!) {
-        deleteCompetition(data:$data)
-    }    
-      `,
-        data: {
-          data: {
-            id: 'fb0b9943-c133-4879-8ff5-1a3c204ed661',
-          },
-        },
-      };
-
-      await supertest(app.getHttpServer())
-        .post('/graphql')
-        .send(postData)
-        .expect(200)
-        .then((response) => {
-          // // Check if data exists in the response body
-          // expect(response.body.data).toBeDefined();
-          // // Check if createCompetition exists in the data
-          // expect(response.body.data.createCompetition).toBeDefined();
-          // // Check if the returned id is correct
-          // expect(response.body.data.createCompetition).toEqual('fb0b9943-c133-4879-8ff5-1a3c204ed661');
-        });
-
-      // retreive competition
-      const query = `
-    query Competition {
-      competition(id: "fb0b9943-c133-4879-8ff5-1a3c204ed661") {
-          id
-          category
-          state
-          deltek_id
-          created_at
-          updated_at
-          deleted_at
-      }
-    } `;
-
-      await supertest(app.getHttpServer())
-        .post('/graphql')
-        .send({ query })
-        .expect(200)
-        .then((response) => {
-          // Check if 'data' and 'competition' are defined
-          expect(response.body.data).toBeDefined();
-          expect(response.body.data.competition).toBeDefined();
-
-          // Check if 'deleted_at' is NOT null
-          expect(response.body.data.competition.deleted_at).toBeDefined();
-        });
-    });
-  });
-
-  // it(`should create application`, async () => {
-  //   const query = `
-  //     mutation CreateApplication($data: CreateApplicationInput!) {
-  //       createApplication(data: $data)
-  //     }
-  //   `;
-
-  //   const variables = {
-  //     data: {
-  //       id: '123e4567-e89b-12d3-a456-426614174010',
-  //       json: { a: 'b' },
-  //     },
-  //   };
-
-  //   return supertest(app.getHttpServer())
-  //     .post('/graphql')
-  //     .send({ query, variables })
-  //     .expect(200)
-  //     .then((response) => {
-  //       console.log('create response: ', response.body);
-
-  //       // // Check that users are defined and an array
-  //       // expect(response.body.data.locations).toBeDefined();
-  //       // expect(Array.isArray(response.body.data.locations)).toBe(true);
-
-  //       // // Check there is only one user
-  //       // expect(response.body.data.locations.length).toEqual(168);
-  //     });
-  // });
+  seedTests();
+  competitionTests();
 
   afterAll(async () => {
     if (app) await app.close();
